WIP: Spacing and Pagebreaks
PDF generates with a wierd break in the bottom page, continuing the text line even after the page is long gone, leaving me to use page breaks. However, there's currently a bug where the page break function does not work as intended, and throws compilation errors. Will be fixed in a future update.

- added more detail and different projects
- added high school education
- changed accent color to green-blue